requirements.txt
watchdog==3.0.0
python-pptx==0.6.23
pdf2image==1.17.0
Pillow==10.2.0
python-docx==1.1.0 